requirements.txt
# === Core Bot ===
discord.py>=2.6.0
python-dotenv>=1.0.0
anthropic>=0.42.0

# === Database ===
aiosqlite>=0.20.0          # async SQLite for the orchestrator

# === Data Science ===
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.14.0
scikit-learn>=1.5.0
shap>=0.46.0
matplotlib>=3.9.0
seaborn>=0.13.0

# === HPC / Remote ===
asyncssh>=2.17.0           # async SSH for Slurm job submission
paramiko>=3.5.0            # fallback SSH library

# === Utilities ===
aiohttp>=3.10.0            # async HTTP (used by discord.py, also useful for webhooks)
pyyaml>=6.0                # config files
